[MIG] Portar campos para classe do product.template
